+++ android/src/main/java/com/crazecoder/flutterbugly/callback/UpgradeCallback.java
@@ -1,0 +1,7 @@
+package com.crazecoder.flutterbugly.callback;
+
+import com.tencent.bugly.beta.UpgradeInfo;
+
+public interface UpgradeCallback {
+    void onUpgrade(UpgradeInfo strategy);
+}